+++ test/types.array.test.js
@@ -37,10 +37,9 @@ module.exports = {
     Array.isArray(a).should.be.true;
     (a._atomics.constructor).should.eql(Object);
   },
-  
+
   'test indexOf()': function(){
     var db = start()
-      , a = new MongooseArray
       , User = db.model('User', 'users_' + random())
       , Pet = db.model('Pet', 'pets' + random());
 
@@ -55,7 +54,7 @@ module.exports = {
     tj.pets.push(jane);
 
     var pending = 3;
-  
+
     [tobi, loki, jane].forEach(function(pet){
       pet.save(function(){
         --pending || done();
@@ -76,5 +75,34 @@ module.exports = {
         });
       });
     }
+  },
+
+  'test #splice()': function () {
+    var collection = 'splicetest' + random();
+    var db = start()
+      , schema = new Schema({ numbers: Array })
+      , A = db.model('splicetest', schema, collection);
+
+    var a = new A({ numbers: [4,5,6,7] });
+    a.save(function (err) {
+      should.equal(null, err, 'could not save splice test');
+      A.findById(a._id, function (err, doc) {
+        should.equal(null, err, 'error finding splice doc');
+        doc.numbers.splice(1, 1);
+        doc.numbers.toObject().should.eql([4,6,7]);
+        doc.save(function (err) {
+          should.equal(null, err, 'could not save splice test');
+          A.findById(a._id, function (err, doc) {
+            should.equal(null, err, 'error finding splice doc');
+            doc.numbers.toObject().should.eql([4,6,7]);
+
+            A.collection.drop(function (err) {
+              db.close();
+              should.strictEqual(err, null);
+            });
+          });
+        });
+      });
+    });
   }
 };